Testing adding cosmiconfig
