+++ packages/jest-config/package.json
@@ -12,7 +12,6 @@
     "babel-jest": "^22.4.1",
     "chalk": "^2.0.1",
     "glob": "^7.1.1",
-    "humps": "^2.0.1",
     "jest-environment-jsdom": "^22.4.1",
     "jest-environment-node": "^22.4.1",
     "jest-get-type": "^22.1.0",
@@ -21,7 +20,7 @@
     "jest-resolve": "^22.4.2",
     "jest-util": "^22.4.1",
     "jest-validate": "^22.4.2",
-    "js-yaml": "^3.11.0",
-    "pretty-format": "^22.4.0"
+    "pretty-format": "^22.4.0",
+    "cosmiconfig": "^4.0.0"
   }
 }
